apply adjustments
