+++ Makefile
@@ -1,12 +1,3 @@
-export PYTHONPATH=$(shell pwd)/hyperon_das_atomdb/
-export PYTHONDONTWRITEBYTECODE=1
-
-test-integration:
-	@pytest -sx ./tests/integration
-
-test-unit:
-	@py.test -sx ./tests/unit --cov=./hyperon_das_atomdb/ --cov-report=term-missing --cov-fail-under=70
-
 isort:
 	@isort ./hyperon_das_atomdb ./tests --multi-line=3 --trailing-comma --force-grid-wrap=0 --use-parentheses --line-width=100
 
@@ -14,6 +5,17 @@ black:
 	@black ./hyperon_das_atomdb ./tests --line-length 100 -t py37 --skip-string-normalization
 
 flake8:
-	@flake8 --show-source ./hyperon_das_atomdb ./tests
+	@flake8 ./hyperon_das_atomdb ./tests --show-source --extend-ignore E501
 
 lint: isort black flake8
+
+unit-tests:
+	@py.test -sx -vv ./tests/unit
+
+unit-tests-coverage:
+	@py.test -sx -vv ./tests/unit --cov=./hyperon_das_atomdb/ --cov-report=term-missing --cov-fail-under=70
+
+integration-tests:
+	@py.test -sx -vv ./tests/integration
+
+pre-commit: unit-tests-coverage lint
