--- hyperon_das_atomdb/constants/redis_mongo_db.py
@@ -1,28 +1,0 @@
-from enum import Enum
-
-
-class MongoCollectionNames(str, Enum):
-    NODES = 'nodes'
-    ATOM_TYPES = 'atom_types'
-    LINKS_ARITY_1 = 'links_1'
-    LINKS_ARITY_2 = 'links_2'
-    LINKS_ARITY_N = 'links_n'
-
-
-class MongoFieldNames(str, Enum):
-    NODE_NAME = 'name'
-    TYPE_NAME = 'named_type'
-    TYPE_NAME_HASH = 'named_type_hash'
-    ID_HASH = '_id'
-    TYPE = 'composite_type_hash'
-    COMPOSITE_TYPE = 'composite_type'
-    KEY_PREFIX = 'key'
-    KEYS = 'keys'
-
-
-class RedisCollectionNames(str, Enum):
-    INCOMING_SET = 'incomming_set'
-    OUTGOING_SET = 'outgoing_set'
-    PATTERNS = 'patterns'
-    TEMPLATES = 'templates'
-    NAMED_ENTITIES = 'names'
