+++ pyproject.toml
@@ -21,6 +21,27 @@ isort = "^5.12.0"
 black = "^23.7.0"
 pytest = "^7.4.2"
 pytest-cov = "^4.1.0"
+flake8-pyproject = "^1.2.3"
+
+
+[tool.isort]
+py_version = 38
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+line_length = 100
+
+
+[tool.black]
+line-length = 100
+target-version = ['py38']
+skip-string-normalization = true
+
+
+[tool.flake8]
+show-source = true
+extend-ignore = ["E501"]
 
 
 [build-system]
